+++ build_ext.py
@@ -14,7 +14,6 @@ except ImportError:
 _LOGGER = logging.getLogger(__name__)
 
 TO_CYTHONIZE = [
-    #    "src/annotatedyaml/objects.py",
     "src/annotatedyaml/reference.py",
 ]
